requirements.txt
numba
numpy
openai
pandas
plotly
protobuf
PyYAML
streamlit
tqdm
tushare
